requirements-dev.txt
flake8
pytest
fastapi
httpx
python-dotenv